Use a process with message passing for otmpd's handler.
Also, changed the protocol up a bit so that the atoms that are
generated are a lot easier to use. Additionally, the 'meta' message
was moved into a otmp_meta_proto protobuf, where we're now using an
enum to denote message type (rather than trusting a string). The gpb
module maps enums to erlang atoms, so this makes it really easy for
us to pattern match on the msgtype in the meta packet.