define protocol packages, as a courtesy
In the protobuf implementation we use, it really doesn't matter,
erl gpb doesn't even use package names by default, we'd have to
enable them. However, if a third party wants to try to build their
own client, it's common practice to define a namespace for your
protocol methods to live in, so they don't clash with anything in
the programmers standard/root namespace.